# docs/swagger-api.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/arc-api/v1/projects/offline/upload-files": {
      "post": {
        "description": "Creates a new project by uploading ACDB and workspace files",
        "operationId": "ProjectController_createProjectFromFiles",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Upload two required files: acdbFile and workspaceFile",
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "acdbFile": {
                    "type": "string",
                    "format": "binary"
                  },
                  "workspaceFile": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "acdbFile",
                  "workspaceFile"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "File opened successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ProjectInfoResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Create project by uploading ACDB and workspace files",
        "tags": [
          "Project"
        ]
      }
    },
...
    "/arc-api/v1/projects/{projectId}": {
      "get": {
        "description": "Get project information based on project Id.",
        "operationId": "ProjectController_getProject",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Id of project",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ProjectInfoResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Project does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Get project information",
        "tags": [
          "Project"
        ]
      },
      "patch": {
        "description": "Update project name and description based on project Id.",
        "operationId": "ProjectController_updateProjectInfo",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Id of project",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectInfoUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ProjectInfoResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid inputs",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Project does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Update project name and description",
        "tags": [
          "Project"
        ]
      },
      "delete": {
        "description": "Delete the project based on project Id.",
        "operationId": "ProjectController_deleteProject",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Id of project",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successfully deleted project",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Project does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Delete project",
        "tags": [
          "Project"
        ]
      }
    },
    "/arc-api/v1/projects/{projectId}/connect": {
      "post": {
        "description": "Establish connection to an existing project for active use.",
        "operationId": "ProjectController_connectToProject",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Id of project",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ProjectInfoResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Project does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Connect to existing project",
        "tags": [
          "Project"
        ]
      }
    },
    "/arc-api/v1/projects/{projectId}/disconnect": {
      "post": {
        "description": "Disconnect from project while keeping it available for future connections.",
        "operationId": "ProjectController_disconnectFromProject",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Id of project",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ProjectInfoResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Project does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Disconnect from project",
        "tags": [
          "Project"
        ]
      }
    },
    "/arc-api/v1/projects/{projectId}/download-files": {
      "get": {
        "description": "Download the ACDB and workspace files based on project Id.",
        "operationId": "ProjectController_downloadArcDbFiles",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Id of project",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/DownloadArcDatabaseFilesResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Project does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Download the ACDB and workspace files",
        "tags": [
          "Project"
        ]
      }
    },
...
  },
  "info": {
    "title": "AudioReach Creator API",
    "description": "Audio Reach Creator API is an open source and cross platform backend framework for updating Audio Reach database files.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter JWT token"
      }
    },
    "schemas": {
...
      "ApiResult": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "success",
          "message"
        ]
      },
      "ProjectInfoResponseDto": {
        "type": "object",
        "properties": {
          "projectId": {
            "type": "string",
            "description": "Unique identifier of the project"
          },
          "name": {
            "type": "string",
            "description": "Human‑readable name of the project"
          },
          "description": {
            "type": "string",
            "description": "Detailed description of the project"
          },
          "projectType": {
            "type": "string",
            "enum": [
              "OFFLINE",
              "DEVICE"
            ],
            "description": "Type of the project (offline or device)"
          },
          "sessionMode": {
            "type": "string",
            "enum": [
              "DIFF_MERGE",
              "DESIGNER",
              "SIMULATION",
              "CONNECTED",
              "DISCONNECTED"
            ],
            "description": "Current session mode for the project"
          }
        },
        "required": [
          "projectId",
          "name",
          "description",
          "projectType",
          "sessionMode"
        ]
      },
      "ProjectInfoUpdateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Optional new name for the project"
          },
          "description": {
            "type": "string",
            "description": "Optional new description for the project"
          },
          "diffMergeDescription": {
            "type": "string",
            "description": "Optional description for a diff‑merge operation"
          }
        }
      },
      "Buffer": {
        "type": "object",
        "properties": {}
      },
      "FileInfo": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the file"
          },
          "fileType": {
            "type": "string",
            "description": "MIME type or file extension"
          },
          "content": {
            "description": "Binary content of the file",
            "allOf": [
              {
                "$ref": "#/components/schemas/Buffer"
              }
            ]
          }
        },
        "required": [
          "name",
          "fileType",
          "content"
        ]
      },
      "DownloadArcDatabaseFilesResponseDto": {
        "type": "object",
        "properties": {
          "acdbFile": {
            "description": "Acdb file information",
            "allOf": [
              {
                "$ref": "#/components/schemas/FileInfo"
              }
            ]
          },
          "workspaceFile": {
            "description": "Workspace file information",
            "allOf": [
              {
                "$ref": "#/components/schemas/FileInfo"
              }
            ]
          }
        },
        "required": [
          "acdbFile",
          "workspaceFile"
        ]
      },
...
    }
  }
}